v0.3

+++ src/lapfarsc/qe/firstrun/dto/AtomDTO.java
@@ -4,9 +4,9 @@ public class AtomDTO {
 
 	private String elementId;
 	private String elementType;
-	private Float x;
-	private Float y;
-	private Float z;
+	private Double x;
+	private Double y;
+	private Double z;
 	
 	private String linha;
 	
@@ -29,23 +29,23 @@ public class AtomDTO {
 	public void setLinha(String linha) {
 		this.linha = linha;
 	}
-	public Float getX() {
+	public Double getX() {
 		return x;
 	}
-	public void setX(Float x) {
+	public void setX(Double x) {
 		this.x = x;
 	}
-	public Float getY() {
+	public Double getY() {
 		return y;
 	}
-	public void setY(Float y) {
+	public void setY(Double y) {
 		this.y = y;
 	}
-	public Float getZ() {
+	public Double getZ() {
 		return z;
 	}
-	public void setZ(Float z) {
+	public void setZ(Double z) {
 		this.z = z;
 	}
-
+	
 }
